commercial/requirements.txt
# ===========================================
# Prometheus-Siren Dependencies
# ===========================================

# --- Core ---
qdrant-client>=1.7.0
google-genai>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# --- Web Framework ---
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0

# --- AST & Code Analysis ---
# (Uses built-in ast module, no extra deps needed)

# --- Docker SDK (for Siren sandbox) ---
docker>=7.0.0

# --- Testing ---
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# --- Logging & Utilities ---
loguru>=0.7.0
rich>=13.7.0
typer>=0.9.0

# --- ML (Hybrid Intelligence) ---
xgboost>=2.0.0
scikit-learn>=1.4.0
transformers>=4.36.0
torch>=2.1.0
jsonpatch>=1.33
